Refactor input for timeout from int to float so we can use fractions of seconds

+++ src/browserist/browser/check_if/is_image_loaded.py
@@ -6,7 +6,7 @@ from ...model.type.xpath import XPath
 from ..wait.for_element import wait_for_element
 
 
-def check_if_is_image_loaded(driver: object, xpath: str, timeout: int = timeout.DEFAULT) -> bool:
+def check_if_is_image_loaded(driver: object, xpath: str, timeout: float = timeout.DEFAULT) -> bool:
     xpath = XPath(xpath)
     wait_for_element(driver, xpath, timeout)
     element: object = driver.find_element(By.XPATH, xpath)  # type: ignore
