+++ src/browserist/browser/get/attribute/__main__.py
@@ -10,7 +10,7 @@ class GetAttributeDriverMethods(DriverMethods):
     def __init__(self, browser_driver: BrowserDriver, settings: BrowserSettings) -> None:
         super().__init__(browser_driver, settings)
 
-    def value(self, xpath: str, attribute: str, timeout: int = timeout.DEFAULT) -> str:
+    def value(self, xpath: str, attribute: str, timeout: float = timeout.DEFAULT) -> str:
         """Get value from an attribute of an element. Examples:
 
         Use "src" as attribute to get the source URL from an <img> image tag.
@@ -19,7 +19,7 @@ class GetAttributeDriverMethods(DriverMethods):
 
         return get_attribute_value(self._driver, xpath, attribute, timeout)
 
-    def values(self, xpath: str, attribute: str, timeout: int = timeout.DEFAULT) -> list[str]:
+    def values(self, xpath: str, attribute: str, timeout: float = timeout.DEFAULT) -> list[str]:
         """Get values from an attribute of multiple elements. Assumes that the XPath targets multiple links. Examples:
 
         Use "src" as attribute to get the source URL from an <img> image tag.
