+++ src/browserist/browser/get/url/__main__.py
@@ -18,26 +18,26 @@ class GetUrlDriverMethods(DriverMethods):
 
         return get_current_url(self._driver)
 
-    def from_image(self, xpath: str, timeout: int = timeout.DEFAULT) -> str:
+    def from_image(self, xpath: str, timeout: float = timeout.DEFAULT) -> str:
         """Get URL source from image, e.g. <img> tag.
 
         This method assumes that the image shouldn't be empty and therefore will retry to get the URL (for better support of single-page apps with extended loading time)."""
 
         return get_url_from_image(self._driver, xpath, timeout)
 
-    def from_images(self, xpath: str, timeout: int = timeout.DEFAULT) -> list[str]:
+    def from_images(self, xpath: str, timeout: float = timeout.DEFAULT) -> list[str]:
         """Get array of URLs from images, e.g. <img> tags. Assumes that the XPath targets multiple images."""
 
         return get_url_from_images(self._driver, xpath, timeout)
 
-    def from_link(self, xpath: str, timeout: int = timeout.DEFAULT) -> str:
+    def from_link(self, xpath: str, timeout: float = timeout.DEFAULT) -> str:
         """Get URL from link, e.g. <a> tag or button.
 
         This method assumes that the link shouldn't be empty and therefore will retry to get the URL (for better support of single-page apps with extended loading time)."""
 
         return get_url_from_link(self._driver, xpath, timeout)
 
-    def from_links(self, xpath: str, timeout: int = timeout.DEFAULT) -> list[str]:
+    def from_links(self, xpath: str, timeout: float = timeout.DEFAULT) -> list[str]:
         """Get array of URLs from links, e.g. <a> tags or buttons. Assumes that the XPath targets multiple links."""
 
         return get_url_from_links(self._driver, xpath, timeout)
