+++ src/browserist/browser/wait/until/number_of_window_handles_is.py
@@ -6,7 +6,7 @@ from ....constant import timeout
 from ....exception.timeout import WaitForWindowTimeoutException
 
 
-def wait_until_number_of_window_handles_is(driver: object, expected_handles: int, timeout: int = timeout.DEFAULT) -> None:
+def wait_until_number_of_window_handles_is(driver: object, expected_handles: int, timeout: float = timeout.DEFAULT) -> None:
     try:
         WebDriverWait(driver, timeout).until(EC.number_of_windows_to_be(expected_handles))  # type: ignore
     except TimeoutException:
