+++ src/browserist/browser/wait/until/text/changes.py
@@ -5,7 +5,7 @@ from ....get.text import get_text
 from ...for_element import wait_for_element
 
 
-def wait_until_text_changes(driver: object, xpath: str, baseline_text: str, timeout: int = timeout.DEFAULT) -> None:
+def wait_until_text_changes(driver: object, xpath: str, baseline_text: str, timeout: float = timeout.DEFAULT) -> None:
     def has_text_changed(driver: object, baseline_text: str) -> bool:
         return get_text(driver, xpath) != baseline_text
