+++ src/browserist/browser/wait/until/text/equals.py
@@ -7,7 +7,7 @@ from ....get.text import get_text
 from ...for_element import wait_for_element
 
 
-def wait_until_text_equals(driver: object, xpath: str, regex: str, timeout: int = timeout.DEFAULT) -> None:
+def wait_until_text_equals(driver: object, xpath: str, regex: str, timeout: float = timeout.DEFAULT) -> None:
     def is_element_text(driver: object, xpath: str, regex: str) -> bool:
         text = get_text(driver, xpath, constant.timeout.BYPASS)
         return bool(re.match(regex, text))
