+++ src/browserist/browser/wait/until/url/changes.py
@@ -4,7 +4,7 @@ from .....model.type.url import URL
 from ....get.url.current import get_current_url
 
 
-def wait_until_url_changes(driver: object, baseline_url: str, timeout: int = timeout.DEFAULT) -> None:
+def wait_until_url_changes(driver: object, baseline_url: str, timeout: float = timeout.DEFAULT) -> None:
     def has_url_changed(driver: object, baseline_url: str) -> bool:
         return get_current_url(driver) != baseline_url
