+++ src/browserist/model/timeout_strategy.py
@@ -18,5 +18,5 @@ class TimeoutStrategy(Enum):
 @dataclass(slots=True)
 class TimeoutSettings:
     strategy: TimeoutStrategy = TimeoutStrategy.BREAK
-    seconds: int = timeout.DEFAULT
+    seconds: float = timeout.DEFAULT
     _is_timed_out: bool = False
